% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{anomaly}
\alias{anomaly}
\title{Unsupervised Anomaly Detection for TabPFN}
\description{
Functions to perform unsupervised anomaly detection using the TabPFN
unsupervised extension from tabpfn_extensions.
}
