% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{augment.tabpfn_classifier}
\alias{augment.tabpfn_classifier}
\title{Augment data with TabPFN classifier predictions}
\usage{
\method{augment}{tabpfn_classifier}(x, new_data, ...)
}
\arguments{
\item{x}{A fitted \code{tabpfn_classifier} object.}

\item{new_data}{A data.frame, tibble, data.table, or matrix.}

\item{...}{Not used.}
}
\value{
A tibble with original data plus \code{.pred_class} and probability columns.
}
\description{
Binds class predictions and probabilities to the input data.
}
