% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{convert_forecasts_to_tibble}
\alias{convert_forecasts_to_tibble}
\title{Convert Python forecasts to R tibble}
\usage{
convert_forecasts_to_tibble(forecasts, item_id_col)
}
\arguments{
\item{forecasts}{Python TimeSeriesDataFrame with forecasts}

\item{item_id_col}{Name of item_id column}
}
\value{
R tibble with forecasts
}
\description{
Convert Python forecasts to R tibble
}
\keyword{internal}
