% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{convert_quantile_predictions}
\alias{convert_quantile_predictions}
\title{Convert TabPFN quantile predictions safely}
\usage{
convert_quantile_predictions(preds_py, quantiles, n_obs)
}
\arguments{
\item{preds_py}{Raw Python predictions object}

\item{quantiles}{Vector of quantiles that were requested}

\item{n_obs}{Number of observations}
}
\value{
Matrix of predictions (n_obs x n_quantiles)
}
\description{
Convert TabPFN quantile predictions safely
}
