% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{create_timeseries_dataframe}
\alias{create_timeseries_dataframe}
\title{Create TimeSeriesDataFrame from R data frame}
\usage{
create_timeseries_dataframe(
  df,
  date_col,
  value_col,
  item_id_col,
  tabpfn_mode = NULL
)
}
\arguments{
\item{df}{R data frame}

\item{date_col}{Name of date column}

\item{value_col}{Name of value column}

\item{item_id_col}{Name of item_id column}
}
\value{
Python TimeSeriesDataFrame object
}
\description{
Create TimeSeriesDataFrame from R data frame
}
\keyword{internal}
