% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_python_object}
\alias{debug_python_object}
\title{Debug Python object structure}
\usage{
debug_python_object(py_obj)
}
\arguments{
\item{py_obj}{A Python object from reticulate}
}
\value{
Prints diagnostic information about the object
}
\description{
Debug Python object structure
}
