% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{extract_tabpfn_engine}
\alias{extract_tabpfn_engine}
\title{Unwrap a parsnip/workflow fit to the TabPFN engine object}
\usage{
extract_tabpfn_engine(object)
}
\description{
Unwrap a parsnip/workflow fit to the TabPFN engine object
}
\keyword{internal}
