% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{find_date_column}
\alias{find_date_column}
\title{Find date/datetime column in data frame}
\usage{
find_date_column(df)
}
\arguments{
\item{df}{Data frame to search}
}
\value{
Name of the first date/datetime column or NULL
}
\description{
Find date/datetime column in data frame
}
\keyword{internal}
