% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{find_value_column}
\alias{find_value_column}
\title{Find value column in data frame}
\usage{
find_value_column(df, date_col, item_id_col)
}
\arguments{
\item{df}{Data frame to search}

\item{date_col}{Name of date column to exclude}

\item{item_id_col}{Name of item_id column to exclude}
}
\value{
Name of the first numeric column or NULL
}
\description{
Find value column in data frame
}
\keyword{internal}
