% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{fit.tab_icl}
\alias{fit.tab_icl}
\title{Fit a TabICL model}
\usage{
\method{fit}{tab_icl}(object, formula = NULL, data = NULL, control = parsnip::control_fit(), ...)
}
\arguments{
\item{object}{A model specification}

\item{formula}{A formula specifying the model}

\item{data}{A data frame}

\item{control}{A `parsnip::control_fit()` object}

\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
Fit a TabICL model
}
