% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{fit.tab_pfn}
\alias{fit.tab_pfn}
\title{Fit a TabPFN model}
\usage{
\method{fit}{tab_pfn}(object, formula = NULL, data = NULL, control = parsnip::control_fit(), ...)
}
\arguments{
\item{object}{A model specification}

\item{formula}{A formula specifying the model}

\item{data}{A data frame}

\item{control}{A `parsnip::control_fit()` object}

\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
Fit a TabPFN model
}
