% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{fit.tab_pfn_ts}
\alias{fit.tab_pfn_ts}
\title{Fit a TabPFN Time Series model}
\usage{
\method{fit}{tab_pfn_ts}(object, formula = NULL, data = NULL, control = parsnip::control_fit(), ...)
}
\arguments{
\item{object}{A model specification}

\item{formula}{A formula specifying the model (optional, columns auto-detected if NULL)}

\item{data}{A data frame}

\item{control}{A `parsnip::control_fit()` object}

\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
Fit a TabPFN Time Series model
}
