% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{fit_tabpfn_ts}
\alias{fit_tabpfn_ts}
\title{Fit TabPFN Time Series model specification (Legacy wrapper)}
\usage{
fit_tabpfn_ts(model_spec, data, ...)
}
\arguments{
\item{model_spec}{A tab_pfn_ts model specification created by [tab_pfn_ts()]}

\item{data}{A data frame containing the training data with date and value columns}

\item{...}{Additional arguments passed to the fitting function}
}
\value{
A fitted model object of class tab_pfn_ts_fit
}
\description{
Fit TabPFN Time Series model specification (Legacy wrapper)
}
