% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{fit_xy.tab_pfn_ts}
\alias{fit_xy.tab_pfn_ts}
\title{Fit a TabPFN Time Series model with xy interface}
\usage{
\method{fit_xy}{tab_pfn_ts}(object, x, y = NULL, control = parsnip::control_fit(), ...)
}
\arguments{
\item{object}{A model specification}

\item{x}{A data frame with training data (must contain date and value columns)}

\item{y}{Not used (value column extracted from x)}

\item{control}{A `parsnip::control_fit()` object}

\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
Fit a TabPFN Time Series model with xy interface
}
