% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{get_encoding.tab_pfn}
\alias{get_encoding.tab_pfn}
\title{Get encoding for tab_pfn model}
\usage{
get_encoding.tab_pfn()
}
\value{
A tibble with encoding information
}
\description{
Get encoding for tab_pfn model
}
\keyword{internal}
