% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{has_method}
\alias{has_method}
\title{Check if object has a method}
\usage{
has_method(object, method)
}
\arguments{
\item{object}{Object to check}

\item{method}{Method name}
}
\value{
Logical
}
\description{
Check if object has a method
}
\keyword{internal}
