% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{infer_frequency}
\alias{infer_frequency}
\title{Infer time series frequency}
\usage{
infer_frequency(dates)
}
\arguments{
\item{dates}{Date/datetime vector}
}
\value{
String representing frequency (compatible with seq() by parameter)
}
\description{
Infer time series frequency
}
\keyword{internal}
