% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{predict.tab_icl_fit}
\alias{predict.tab_icl_fit}
\title{Make predictions from a fitted TabICL model}
\usage{
\method{predict}{tab_icl_fit}(object, new_data, type = NULL, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{new_data}{A data frame of predictors}

\item{type}{A single character string for the prediction type}

\item{...}{Additional arguments}
}
\value{
A tibble of predictions
}
\description{
Make predictions from a fitted TabICL model
}
\keyword{internal}
