% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{predict.tab_pfn_ts_fit}
\alias{predict.tab_pfn_ts_fit}
\title{Make predictions from a fitted TabPFN Time Series model}
\usage{
\method{predict}{tab_pfn_ts_fit}(object, new_data, type = NULL, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{new_data}{A data frame of time series data}

\item{type}{A single character string for the prediction type}

\item{...}{Additional arguments}
}
\value{
A tibble of forecasts
}
\description{
Make predictions from a fitted TabPFN Time Series model
}
\keyword{internal}
