% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap.R
\name{print.shap_explanation}
\alias{print.shap_explanation}
\title{Print method for SHAP explanation}
\usage{
\method{print}{shap_explanation}(x, ...)
}
\arguments{
\item{x}{A shap_explanation object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for SHAP explanation
}
