% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabicl.R
\name{print.tab_icl}
\alias{print.tab_icl}
\title{Print method for TabICL models}
\usage{
\method{print}{tab_icl}(x, ...)
}
\arguments{
\item{x}{A tab_icl model object or specification}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for TabICL models
}
