% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{print.tab_icl_fit}
\alias{print.tab_icl_fit}
\title{Print method for fitted TabICL model}
\usage{
\method{print}{tab_icl_fit}(x, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{...}{Additional arguments (not used)}
}
\value{
The model object (invisibly)
}
\description{
Print method for fitted TabICL model
}
