% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{print.tab_pfn_fit}
\alias{print.tab_pfn_fit}
\title{Print method for fitted TabPFN model}
\usage{
\method{print}{tab_pfn_fit}(x, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{...}{Additional arguments (not used)}
}
\value{
The model object (invisibly)
}
\description{
Print method for fitted TabPFN model
}
