% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{print.tab_pfn_time_series}
\alias{print.tab_pfn_time_series}
\title{Print method for TabPFN Time Series models}
\usage{
\method{print}{tab_pfn_time_series}(x, ...)
}
\arguments{
\item{x}{A tab_pfn_time_series model object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for TabPFN Time Series models
}
