% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{print.tab_pfn_ts}
\alias{print.tab_pfn_ts}
\title{Print method for TabPFN Time Series model specification}
\usage{
\method{print}{tab_pfn_ts}(x, ...)
}
\arguments{
\item{x}{A model specification}

\item{...}{Additional arguments (not used)}
}
\value{
The model specification (invisibly)
}
\description{
Prints the model specification details for a TabPFN Time Series model.
}
