% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{print.tab_pfn_ts_fit}
\alias{print.tab_pfn_ts_fit}
\title{Print method for fitted TabPFN Time Series model}
\usage{
\method{print}{tab_pfn_ts_fit}(x, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{...}{Additional arguments (not used)}
}
\value{
The model object (invisibly)
}
\description{
Prints details of a fitted TabPFN Time Series model.
}
