% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{print.tab_pfn_unsupervised}
\alias{print.tab_pfn_unsupervised}
\title{Print method for TabPFN Unsupervised models}
\usage{
\method{print}{tab_pfn_unsupervised}(x, ...)
}
\arguments{
\item{x}{A tab_pfn_unsupervised model object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for TabPFN Unsupervised models
}
