% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-embeddings.R
\name{reduce_dims}
\alias{reduce_dims}
\title{Run t-SNE or UMAP on an embedding matrix}
\usage{
reduce_dims(embeddings, method)
}
\description{
Run t-SNE or UMAP on an embedding matrix
}
\keyword{internal}
