% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{register_tabpfn_parsnip}
\alias{register_tabpfn_parsnip}
\title{Register TabPFN as a parsnip engine}
\usage{
register_tabpfn_parsnip()
}
\description{
Register TabPFN as a parsnip engine
}
\keyword{internal}
