% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{required_pkgs.tab_pfn}
\alias{required_pkgs.tab_pfn}
\title{Check that the required arguments are available for fitting}
\usage{
required_pkgs.tab_pfn(object, ...)
}
\arguments{
\item{object}{A model specification}

\item{x}{A data frame or matrix of predictors}
}
\value{
TRUE
}
\description{
Check that the required arguments are available for fitting
}
\keyword{internal}
