% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{set_engine.tab_icl}
\alias{set_engine.tab_icl}
\title{Set the model engine for TabICL}
\usage{
set_engine.tab_icl(x, eng = c("tabicl"), ...)
}
\arguments{
\item{x}{A model specification}

\item{eng}{A single character string for the engine (always "tabicl" for this model)}

\item{...}{Additional engine-specific arguments}
}
\value{
An updated model specification
}
\description{
Set the model engine for TabICL
}
