% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{set_engine.tab_pfn}
\alias{set_engine.tab_pfn}
\title{Set the model engine for TabPFN}
\usage{
set_engine.tab_pfn(x, eng = c("tabpfn"), ...)
}
\arguments{
\item{x}{A model specification}

\item{eng}{A single character string for the engine (always "tabpfn" for this model)}

\item{...}{Additional engine-specific arguments}
}
\value{
An updated model specification
}
\description{
Set the model engine for TabPFN
}
