% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{set_engine.tab_pfn_ts}
\alias{set_engine.tab_pfn_ts}
\title{Set the model engine for TabPFN Time Series}
\usage{
set_engine.tab_pfn_ts(x, eng = c("tabpfn_ts"), ...)
}
\arguments{
\item{x}{A model specification}

\item{eng}{A single character string for the engine (always "tabpfn_ts")}

\item{...}{Additional engine-specific arguments}
}
\value{
An updated model specification
}
\description{
Set the model engine for TabPFN Time Series
}
