% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{set_mode.tab_icl}
\alias{set_mode.tab_icl}
\title{Set mode for TabICL}
\usage{
set_mode.tab_icl(object, mode, ...)
}
\arguments{
\item{object}{A model specification}

\item{mode}{A single character string for prediction mode}

\item{...}{Not used}
}
\value{
An updated model specification
}
\description{
Set mode for TabICL
}
