% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabpfn.R
\name{set_mode.tab_pfn}
\alias{set_mode.tab_pfn}
\title{Set mode for TabPFN}
\usage{
set_mode.tab_pfn(object, mode, ...)
}
\arguments{
\item{object}{A model specification}

\item{mode}{A single character string for prediction mode}

\item{...}{Not used}
}
\value{
An updated model specification
}
\description{
Set mode for TabPFN
}
