% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-timeseries.R
\name{set_mode.tab_pfn_ts}
\alias{set_mode.tab_pfn_ts}
\title{Set mode for TabPFN Time Series}
\usage{
set_mode.tab_pfn_ts(object, mode, ...)
}
\arguments{
\item{object}{A model specification}

\item{mode}{A single character string for prediction mode}

\item{...}{Not used}
}
\value{
An updated model specification
}
\description{
Set mode for TabPFN Time Series
}
