% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_classifier_inputs}
\alias{validate_classifier_inputs}
\title{Validate classifier inputs}
\usage{
validate_classifier_inputs(x, y)
}
\arguments{
\item{x}{A data.frame, tibble, data.table, or matrix.}

\item{y}{A factor, character, or integer vector.}
}
\value{
A factor vector (validated and possibly coerced).
}
\description{
Validate classifier inputs
}
\keyword{internal}
