% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_regressor_inputs}
\alias{validate_regressor_inputs}
\title{Validate regressor inputs}
\usage{
validate_regressor_inputs(x, y)
}
\arguments{
\item{x}{A data.frame, tibble, data.table, or matrix.}

\item{y}{A numeric vector.}
}
\value{
A double vector (validated).
}
\description{
Validate regressor inputs
}
\keyword{internal}
