% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-tabicl.R
\name{fit_xy.tab_icl}
\alias{fit_xy.tab_icl}
\title{Fit a TabICL model with xy interface}
\usage{
fit_xy.tab_icl(object, x, y, control = parsnip::control_fit(), ...)
}
\arguments{
\item{object}{A model specification}

\item{x}{A data frame or matrix of predictors}

\item{y}{A vector of outcomes}

\item{control}{A `parsnip::control_fit()` object}

\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
Fit a TabICL model with xy interface
}
