% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabpfn.R
\name{print.tab_pfn}
\alias{print.tab_pfn}
\title{Print method for TabPFN models}
\usage{
\method{print}{tab_pfn}(x, ...)
}
\arguments{
\item{x}{A tab_pfn model object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for TabPFN models
}
