% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtabpfn-package.R
\name{setup_tabpfn}
\alias{setup_tabpfn}
\title{Configure TabPFN Python Environment}
\usage{
setup_tabpfn(envname = "r-tabpfn", force = FALSE, install_shap = FALSE)
}
\arguments{
\item{envname}{Name of the virtual environment}

\item{force}{Logical. If TRUE, recreates environment even if it exists}

\item{install_shap}{Logical. If TRUE, installs tabpfn-extensions for SHAP support}
}
\value{
NULL (invisible)
}
\description{
Sets up of Python environment for TabPFN usage.
}
\examples{
\dontrun{
# Setup environment
setup_tabpfn()

# Setup with SHAP support
setup_tabpfn(install_shap = TRUE)
}
}
