% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtabpfn-package.R
\name{setup_tabpfn}
\alias{setup_tabpfn}
\title{Configure TabPFN Python Environment}
\usage{
setup_tabpfn(envname = "r-tabpfn", force = FALSE)
}
\arguments{
\item{envname}{Name of the virtual environment}

\item{force}{Logical. If TRUE, recreates the environment even if it exists}
}
\value{
NULL (invisible)
}
\description{
Sets up the Python environment for TabPFN usage.
}
\examples{
\dontrun{
# Setup environment
setup_tabpfn()
}
}
