% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabpfn.R
\name{tab_pfn_classification}
\alias{tab_pfn_classification}
\title{Train a TabPFN classification model}
\usage{
tab_pfn_classification(X, y, device = "auto", test_size = 0.33, ...)
}
\arguments{
\item{X}{Predictor data frame or matrix}

\item{y}{Response vector (factor or character)}

\item{device}{Device to use: "auto", "cpu", or "cuda"}

\item{test_size}{Proportion of data to use for internal validation}

\item{...}{Additional arguments passed to TabPFNClassifier}
}
\value{
A tab_pfn model object with mode = "classification"
}
\description{
Train a TabPFN classification model
}
